+++ src/Sql/Sql.php
@@ -221,23 +221,66 @@ class Sql {
     
     // Connexion par defaut ?
     $this->is_default = $is_default;
+
     // Définition des données de connexion
-    $this->db_type = isset($db['phptype']) ? $db['phptype'] : 'pgsql';
-    $this->cnxstring = "$this->db_type:dbname=$db[database];host=$db[hostspec];port=$db[port]";
-    $this->username = $db['username'];
-    $this->password = $db['password'];
-    $this->persistent = $db['persistent'];
+    $this->username = isset($db['username']) ? $db['username'] : null;
+    $this->password = isset($db['password']) ? $db['password'] : null;
+    $this->persistent = isset($db['persistent']) ? $db['persistent'] : false;
+
+    // MANTIS 0007565: [SQL] Gérer la connexion au format DSN
+    if (isset($db['dsn'])) {
+      $this->connectionFromDSN($db);
+    }
+    else {
+      $this->db_type = isset($db['phptype']) ? $db['phptype'] : 'pgsql';
+      $this->cnxstring = "$this->db_type:dbname=$db[database];host=$db[hostspec];port=$db[port]";
+    }
+
     // Définition des données de connexion pour la lecture
     if (isset($db_read)) {
-      $this->cnxstring_read = "$this->db_type:dbname=$db_read[database];host=$db_read[hostspec];port=$db_read[port]";
-      $this->username_read = $db_read['username'];
-      $this->password_read = $db_read['password'];
-      $this->persistent_read = $db_read['persistent'];
+      $this->username_read = isset($db_read['username']) ? $db_read['username'] : null;
+      $this->password_read = isset($db_read['password']) ? $db_read['password'] : null;
+      $this->persistent_read = isset($db_read['persistent']) ? $db_read['persistent'] : false;
+
+      // MANTIS 0007565: [SQL] Gérer la connexion au format DSN
+      if (isset($db_read['dsn'])) {
+        $this->connectionFromDSN($db_read, true);
+      }
+      else {
+        $this->cnxstring_read = "$this->db_type:dbname=$db_read[database];host=$db_read[hostspec];port=$db_read[port]";
+      }
+      
     }
     // Mise en cache des statements
     // MANTIS 3547: Réutiliser les prepare statements pour les requêtes identiques
     $this->PreparedStatementCache = [];
     $this->getConnection();
+  }
+
+  /**
+   * Gerer une connexion à partir du DSN
+   * 
+   * @param array $db
+   * @param boolean $isread
+   */
+  protected function connectionFromDSN($db, $isread = false) {
+    // Gérer le type de connexion
+    if (strpos($db['dsn'], ':') === false) {
+      $dsn = explode(';', $db['dsn']);
+      $this->db_type = 'pgsql';
+    }
+    else {
+      $dsn = explode(':', $db['dsn'], 2);
+      $this->db_type = $dsn[0];
+      $dsn = explode(';', $dsn[1]);
+    }
+    
+    if ($isread) {
+      $this->cnxstring_read = $db['dsn'];
+    }
+    else {
+      $this->cnxstring = $db['dsn'];
+    }
   }
 
   /**